feat: add Logger with a bunch of tests
This allows to log all the freaking requests received.
HELL YEAH.
You can simply use it as is, all the doc is in the README.md.
Some tests are written in order to check if it behaves as expected, the
there is no verification, it just passes the tests and gives the output
of requests received on multiple pipelines.